Add support for accessing multiple API base paths
With the introduction of the Hetzner API (currently only for storage
boxes) alongside the Hetzner Cloud API, being able to access multiple
API base paths for different API endpoints will become necessary.
This introduces the corresponding infrastructure in the configuration
and API functions for this to work.
Through the configuration's `base_path_mapping` parameter, the user is
still able to overwrite all API base paths, e.g. for mocking the API in
a testing environment.